chore(netlify): unset CI variable in netlify.toml
tl;dr
Netlify deploy error: Treating warnings as errors because of
process.env.CI = true

Netlify started a gradual rollout of adding the environment variable CI
to build environments, with the value of true. This environment
variable, short for Continuous Integration, is commonly set in various
CI environments.  The ecosystem has largely agreed to use this
environment setting to detect when a build is executing in a CI
environment, as opposed to a local development environment